papers/requirements.txt
sentence-transformers>=3.0.0
torch
psycopg2-binary
pgvector
python-dotenv
tqdm
kaggle
einops
